Phase 4c: extract NervesHub.Devices.Health sub-context
Third step of splitting the Devices god-module. Built on Phase 4b
(cleanup/phase-4b-devices-pinning).

Moves the device-health functions (save_device_health/1, truncate_device_health/0,
health_status_count/2, and the private update_health_on_device/1 helper) out of
NervesHub.Devices into a new, documented NervesHub.Devices.Health module. The
now-unused DeviceHealth alias is dropped from Devices.

Function names unchanged; call sites across lib and test repointed to the new
module (dropping now-unused `Devices` aliases where applicable).

Also adds a `.dialyzer_ignore.exs` entry for the new module path: save_device_health
uses an Ecto.Multi chain, which triggers the same known-false-positive
:call_without_opaque warning already ignored for devices.ex (the ignore file is
path-based, so the moved code needs its own entry).

Verification: mix format, credo, dialyzer (0 unskipped), full test suite
(1396 passing; 2 unrelated ConnectionHistoryTest failures were a DBConnection
sandbox-ownership flake that passes in isolation).
