+++ .dialyzer_ignore.exs
@@ -8,6 +8,7 @@
   {"lib/nerves_hub/accounts.ex", :call_without_opaque},
   {"lib/nerves_hub/accounts/remove_account.ex", :call_without_opaque},
   {"lib/nerves_hub/devices.ex", :call_without_opaque},
+  {"lib/nerves_hub/devices/health.ex", :call_without_opaque},
   {"lib/nerves_hub_web/components/layouts.ex", :call},
   {"lib/nerves_hub_web/components/layouts.ex", :no_return}
 ]
